- added .deb package generation to cmake file

+++ inc/executor.h
@@ -157,9 +157,9 @@ public:
 	inline int FindConstant( const Object & o )
 	{
 		if( constants_set.count( o ) == 0 ) return -1;
-		for( int i = 0; i < constants.size(); i++ )
+		for( size_t i = 0; i < constants.size(); i++ )
 			if( o == constants.at( i ) )
-				return i;
+				return (int)i;
 		return -1;
 	}
 	inline Object GetConstant( int idx ) { return constants.at( idx ); }
